Flesh out the FileManipulator interface.
I'm on the fence about the interface to declare for the f value in
FileManipulator. I want this to be a type that can support both a
bytes.Buffer and an os.File. bytes.Buffer does not implement Seek(). I
may replace this with io.ReadWriter and just use whole-buffer
copy/replace instead of seeking.