Passing tests, details below.
The test failures were caused by using Truncate() without a Seek() call.
Per the official docs, Truncate() changes a file's contents without
changing the location of the 'cursor' within that file. The previous
code was truncating and then writing at an offset, causing null bytes
and test failures.

By using Truncate, WriteAt, and Seek in that order I'm able to copy the
file contents over and leave the file in a state fit for manipulation by
the rest of the application.